requirements.txt
astroid==1.6.1
atomicwrites==1.3.0
attrs==19.3.0
boto3==1.6.21
botocore==1.9.21
certifi==2017.11.5
cffi==1.13.1
chardet==3.0.4
Click==7.0
colorama==0.3.9
cryptography==2.8
docutils==0.14
Flask==1.1.1
flask-redis==0.4.0
Flask-SQLAlchemy==2.4.1
idna==2.6
importlib-metadata==0.23
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.3
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
marshmallow==3.2.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
pony==0.7.10
py==1.8.0
py4j==0.10.6
pycparser==2.19
PyJWT==1.7.1
pylint==1.8.2
pyparsing==2.4.4
pyspark==2.3.0
pytest==5.2.2
python-dateutil==2.6.1
python-dotenv==0.10.3
redis==3.3.11
requests==2.18.4
s3transfer==0.1.13
six==1.11.0
SQLAlchemy==1.3.10
SQLAlchemy-Utils==0.34.2
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.10.11
zipp==0.6.0
